Bazel: Fix relative go to definition in external repositories
The resolve root was calculated incorrectly for relative load statements in external repositories. Instead of removing the file's path relative to the external repository, it only kept this part. For example if the current file is `$(bazel info output_base)/external/foo/bar/BUILD`, it would compute the resolve root as `bar/BUILD`, whereas it needs to be `$(bazel info output_base)/external/foo`.