fix: Offset of bare completions
When completing an empty string, the completions for files would include the wrong offset, causing vscode to chop off the end of the line after accepting the completion. This is a relatively simple fix in `get_filesystem_entries`, but a lot of this PR is adding testing and extracting `bazel query` into the `MockBazel` client.

This also introduces a builder pattern for the test context, to allow adding certain information such as the mock results to `bazel query`.