feat: Add autocomplete for bazel builtins
This implements the function `get_environment` on `LspContext`. This reads the a protobuf file generated by bazel that contains all the builtins. This does not include things like `cc_library`. Further investigation is needed on how to obtain these. Type information could be generated better too, but this will happen in a follow-up.