fix: lint for undefined global symbols
Previous implementation was passing empty prelude list to lint.
The starlark-rust does not create warnings for undefined global
symbols in such case.

In general, for bazel we do not have prelude, just a list of global
symbols. Needed to add manually some missing symbols, because bazel does
not report all global symbols in `bazel info build-language`, see
https://github.com/bazel-contrib/vscode-bazel/issues/1#issuecomment-2036369868

In the process:
 - Removed unused prelude. Looks it was carried over with migrating from
   starlark-rust, but prelude is not used in case of bazel.
 - Removed ContextMode run - bazel lsp is not able to actually run any
   starlark, those seems to be leftovers from starlark-rust